requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-silk==5.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
environs==14.3.0
gprof2dot==2025.4.14
greenlet==3.2.4
inflection==0.5.1
marshmallow==4.0.1
packaging==25.0
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
query-counter==0.3.1
redis==6.4.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
